ItemManagerGUI Sort by new Combination
To show the list of combinations in ItemManager you create an object with ItemManagerGUI. In the inspector you can fill up combinationNew once it is full it will automatically be added to the ItemManger combinations list. When you fill up the combinationNew, ItemManagerGUI will only display the combinations that match with it.

So if you search with "flower + none = none", a combination of "flower + character = change color" and any others with flower in the front will show up. While it will not show up for "character + none = none".

Note you can change a field back to none by clicking the small dot in circle thing.