bugfix drag and drop
items wouldn't be drag and droppable after the previous change where the trigger = true was removed. I've put it back and instead made it so that player would extend from a different class collidable item where collidable item inherits from item and just makes sure that there are two boxcollider2d, one is a trigger and the other isn't.